Changelog category (leave one):
- Performance Improvement


Changelog entry (up to few sentences, required except for Non-significant/Documentation categories):
Improved performance of sorting without a limit or with big limit and external sorting.