@alexey-milovidov 2020-01-07
```
SET max_memory_usage = 100000000000, max_bytes_before_external_sort = 50000000000;

SELECT number FROM numbers_mt(1000000) ORDER BY intHash64(number) LIMIT 900000, 10;
old: 0.149 sec.
new: 0.119 sec.

SELECT number FROM numbers_mt(10000000) ORDER BY intHash64(number) LIMIT 9000000, 10;
old: 2.622 sec.
new: 1.218 sec.

SELECT number FROM numbers_mt(100000000) ORDER BY intHash64(number) LIMIT 90000000, 10;
old: 60.771 sec.
new: 11.946 sec.

SELECT number FROM numbers_mt(1000000000) ORDER BY intHash64(number) LIMIT 900000000, 10;
old: 1376.686 sec.
new: 122.873 sec

SELECT number FROM numbers_mt(10000000000) ORDER BY intHash64(number) LIMIT 9000000000, 10;
old: 20719.163 sec.
new: 1636.834 sec
```

Up to 12 times faster on 16 core server.