@alexey-milovidov 2020-01-07
Performance test shows significant improvement:
https://clickhouse-test-reports.s3.yandex.net/8545/2314bb5a3430c421ddc371daeb9d20134bdaf181/performance_test/comparison_min_time_gcc_9.html