@zhang2014 2020-01-07
It changes the result of the following statement(Is that expected?)
```
SELECT * FROM (SELECT a, runningDifference(a) rd FROM (SELECT * FROM (SELECT 1 a UNION ALL SELECT 2 a ) ORDER BY a DESC ));
```