@alexey-milovidov 2020-01-07
It's not guaranteed how the returned data will be split in blocks.
Yes, this result is valid.