@GPF199541 2020-09-27
Connecting to 127.0.0.1:9000 as user default.
Connected to ClickHouse server version 20.9.2 revision 54439.

dbrtmappredb01 :)  SELECT * FROM (SELECT a, runningDifference(a) rd FROM (SELECT * FROM (SELECT 1 a UNION ALL SELECT 2 a ) ORDER BY a DESC ));

SELECT *
FROM 
(
    SELECT 
        a,
        runningDifference(a) AS rd
    FROM 
    (
        SELECT *
        FROM 
        (
            SELECT 1 AS a
            UNION ALL
            SELECT 2 AS a
        )
        ORDER BY a DESC
    )
)

┌─a─┬─rd─┐
│ 2 │  0 │
└───┴────┘
┌─a─┬─rd─┐
│ 1 │  0 │
└───┴────┘

2 rows in set. Elapsed: 0.042 sec. 


the version 20.9.2 is the same question,how to solve this problem