@den-crane 2020-09-27
@GPF199541  No solution. 2 blocks. runningDifference works only in block.
You can use arrays 
```
SELECT arrayJoin(arrayDifference(groupArray(a))) from 
  (select * FROM (SELECT 1 a UNION ALL SELECT 2 a ) ORDER BY a DESC)
```